[680] - setting up pg type basis

+++ diesel/src/pg/types/money.rs
@@ -7,14 +7,14 @@ use types::{self, ToSql, IsNull, FromSql};
 
 primitive_impls!(Oid -> (u32, pg: (26, 1018)));
 
-impl FromSql<types::Oid, Pg> for i64 {
+impl FromSql<types::Oid, Pg> for u32 {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         let mut bytes = not_none!(bytes);
         bytes.read_u32::<NetworkEndian>().map_err(|e| e.into())
     }
 }
 
-impl ToSql<types::Oid, Pg> for i64 {
+impl ToSql<types::Oid, Pg> for u32 {
     fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         out.write_u32::<NetworkEndian>(*self)
             .map(|_| IsNull::No)
